# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BloodLink - Blood Bank Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: List donors
      parameters:
        - name: cin
          in: query
          schema:
            type: string
        - name: bloodType
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of donors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorListResponse"
  /donors/{id}:
    get:
      operationId: getDonor
      tags: [donors]
      summary: Get donor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
  /donors/by-cin/{cin}:
    get:
      operationId: getDonorByCin
      tags: [donors]
      summary: Get donor by CIN
      parameters:
        - name: cin
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Donor found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "404":
          description: Donor not found
  /donors/register:
    post:
      operationId: registerDonor
      tags: [donors]
      summary: Register a donor manually (by center)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDonorBody"
      responses:
        "201":
          description: Donor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
  /donors/{id}/eligibility:
    get:
      operationId: checkDonorEligibility
      tags: [donors]
      summary: Check donor eligibility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"
  /donors/{id}/medical-form:
    post:
      operationId: submitMedicalForm
      tags: [donors]
      summary: Submit pre-donation medical form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalFormBody"
      responses:
        "200":
          description: Form submitted, eligibility evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"

  # APPOINTMENTS
...
  /requests:
    get:
      operationId: listBloodRequests
      tags: [requests]
      summary: List blood requests
      parameters:
        - name: establishmentId
          in: query
          schema:
            type: integer
        - name: centerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [submitted, processing, accepted, shipped, delivered, rejected]
        - name: urgency
          in: query
          schema:
            type: string
            enum: [normal, urgent, critical]
      responses:
        "200":
          description: Blood requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequestListResponse"

    post:
      operationId: createBloodRequest
      tags: [requests]
      summary: Create a blood request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBloodRequestBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
  /requests/{id}:
    get:
      operationId: getBloodRequest
      tags: [requests]
      summary: Get blood request details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

    patch:
      operationId: updateBloodRequest
      tags: [requests]
      summary: Update blood request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBloodRequestBody"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # ALERTS
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: establishmentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, rejected]
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
  /invoices/{id}/pay:
    post:
      operationId: payInvoice
      tags: [invoices]
      summary: Pay an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceBody"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [shortage_alert, reminder, general]
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # DASHBOARD SUMMARIES
...
components:
  schemas:
...
    Donor:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        cin:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        weight:
          type: number
        phone:
          type: string
        email:
          type: string
        region:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        eligibilityStatus:
          type: string
          enum: [eligible, temporarily_excluded, permanently_ineligible]
        totalDonations:
          type: integer
        lastDonationDate:
          type: string
        nextEligibleDate:
          type: string
        annualQuota:
          type: integer
        annualDonationsCount:
          type: integer
      required: [id, firstName, lastName, cin, bloodType, eligibilityStatus]
    DonorListResponse:
      type: object
      properties:
        donors:
          type: array
          items:
            $ref: "#/components/schemas/Donor"
        total:
          type: integer
      required: [donors, total]
    RegisterDonorBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        cin:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        weight:
          type: number
        phone:
          type: string
        email:
          type: string
        region:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
      required: [firstName, lastName, cin, dateOfBirth, gender, weight, bloodType]
    EligibilityResult:
      type: object
      properties:
        eligible:
          type: boolean
        status:
          type: string
          enum: [eligible, temporarily_excluded, permanently_ineligible]
        reason:
          type: string
        nextEligibleDate:
          type: string
      required: [eligible, status]
    MedicalFormBody:
      type: object
      properties:
        currentMedications:
          type: string
        recentSurgeries:
          type: boolean
        recentTravelEndemicZones:
          type: boolean
        recentDentalCare:
          type: boolean
        lastDonationDate:
          type: string
        otherNotes:
          type: string
      required: [currentMedications, recentSurgeries, recentTravelEndemicZones, recentDentalCare]

    # APPOINTMENTS
...
    BloodRequest:
      type: object
      properties:
        id:
          type: integer
        establishmentId:
          type: integer
        establishmentName:
          type: string
        centerId:
          type: integer
        centerName:
          type: string
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        volume:
          type: number
        urgency:
          type: string
          enum: [normal, urgent, critical]
        status:
          type: string
          enum: [submitted, processing, accepted, shipped, delivered, rejected]
        bagIds:
          type: array
          items:
            type: integer
        estimatedDelivery:
          type: string
        submittedAt:
          type: string
        updatedAt:
          type: string
      required: [id, establishmentId, bloodType, volume, urgency, status, submittedAt]
    BloodRequestListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/BloodRequest"
        total:
          type: integer
      required: [requests, total]
    CreateBloodRequestBody:
      type: object
      properties:
        bloodType:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        volume:
          type: number
        urgency:
          type: string
          enum: [normal, urgent, critical]
      required: [bloodType, volume, urgency]
    UpdateBloodRequestBody:
      type: object
      properties:
        status:
          type: string
          enum: [processing, accepted, shipped, delivered, rejected]
        bagIds:
          type: array
          items:
            type: integer
        estimatedDelivery:
          type: string
        rejectionReason:
          type: string
      required: [status]

    # ALERTS
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        establishmentId:
          type: integer
        establishmentName:
          type: string
        requestId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, rejected]
        isEmergency:
          type: boolean
        dueDate:
          type: string
        paidAt:
          type: string
        createdAt:
          type: string
      required: [id, establishmentId, amount, status, createdAt]
    InvoiceListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
      required: [invoices, total]
    PayInvoiceBody:
      type: object
      properties:
        cardNumber:
          type: string
        expiryDate:
          type: string
        cvv:
          type: string
      required: [cardNumber, expiryDate, cvv]

    # NOTIFICATIONS
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [shortage_alert, reminder, general, appointment_update, request_update]
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
        relatedId:
          type: integer
      required: [id, userId, type, title, message, read, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
        total:
          type: integer
      required: [notifications, unreadCount, total]

    # DASHBOARD
...